requirements.txt
flask==3.0.0
google-genai==0.7.0
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.20.0
safetensors>=0.3.0
pillow>=10.0.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3
requests>=2.31.0
pyngrok>=7.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
